split out codegen
also add clippy config and make clippy pass
