+++ clippy.toml
@@ -1,0 +1,1 @@
+allow-unwrap-in-tests = true
